pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verify-ai"
version = "0.1.0"
description = "AI-powered automated verification system for code quality"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "VerifyAI Team" }
]
keywords = ["testing", "ai", "automation", "code-quality", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.0",
    "gitpython>=3.1.0",
    "anthropic>=0.18.0",
    "openai>=1.10.0",
    "ollama>=0.1.0",
    "openapi-spec-validator>=0.7.0",
    "prance>=23.0.0",
    # Phase 2: tree-sitter for multi-language AST parsing
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    "tree-sitter-typescript>=0.21.0",
    "tree-sitter-go>=0.21.0",
    "tree-sitter-java>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
]
tui = [
    "textual>=0.50.0",
]
all = [
    "verify-ai[dev,server,tui]",
]

[project.scripts]
vai = "verify_ai.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/verify-ai"
Documentation = "https://github.com/your-org/verify-ai#readme"
Repository = "https://github.com/your-org/verify-ai"

[tool.hatch.build.targets.wheel]
packages = ["src/verify_ai"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
